+++ Problems/Problem9.cs
@@ -13,22 +13,22 @@ namespace ProjectEuler.Problems
   class Problem9: ProblemBase
   {
     private const double EPSILON = 1e-6;
-    public decimal Solve()
+    public ulong Solve()
     {
-      decimal aValue = 0, bValue = 0, cValue = 0;
+      int aValue = 0, bValue = 0, cValue = 0;
       for (int a = 1; a < 1000; ++a )
       {
         double b = (double)(1000*1000 - 2000*a)/(double)(2000 - 2*a);
         if (Math.Abs(Math.Floor(b) - b) < EPSILON)
         {
           aValue = a;
-          bValue = (decimal)Math.Floor(b);
+          bValue = (int)Math.Floor(b);
           break;
         }
       }
       cValue = 1000 - (aValue + bValue);
 
-      return aValue*bValue*cValue;
+      return (ulong)(aValue * bValue * cValue);
     }
   }
 }
